fix: javax->jakarta로 변경 및 Integer로 변경

+++ src/main/java/com/trianglechoke/codesparring/membercode/entity/MemberCode.java
@@ -35,6 +35,6 @@ public class MemberCode {
   private Quiz quiz;
   
   @Column(name = "quiz_correct", nullable = false, length = 1)
-  private Long quizCorrect;
+  private Integer quizCorrect;
 
 }
